Move Clang Functions from util
