+++ TAIA-test-automation-for-incremental-analysis/scripts/generate_ai_mutations.py
@@ -4,6 +4,7 @@ import random
 import time
 from concurrent.futures import ThreadPoolExecutor
 from multiprocessing import Lock
+from generate_clang_mutations import get_operator_descriptions_for_ai
 
 import openai
 
@@ -154,7 +155,7 @@ def _make_gpt_request(snippet, ai_16k):
     prompt = f'''
         You are a developer for C helping me with my following question. I want to understand the typical process of code evolution by looking at how developers make changes over time for testing an incremental analysis of the static c analyzer Goblint.
 
-        The following mutations are already generated by me. So please do not generate programs that can be generated by this mutations: {getOperatorDescriptionsForAI()}. Please do not consider these mutations as examples for how your code changes should look like. Just try to prevent doing things that could be done with these mutations.
+        The following mutations are already generated by me. So please do not generate programs that can be generated by this mutations: {get_operator_descriptions_for_ai()}. Please do not consider these mutations as examples for how your code changes should look like. Just try to prevent doing things that could be done with these mutations.
 
         Below is an snippet from a C file which represents a part of the finished program. My question is how a previous version of this code could have looked like before some typical code changes done by developers. Please generate me such a previous version!
